+++ Controls-demo/Menu/Control/HoverBackgroundStyle/Index.wml
@@ -4,9 +4,7 @@
             source="{{_source}}"
             keyProperty="key"
             displayProperty="title"
-            itemActions="{{_itemActions}}"
             attr:class="ControlsDemo-menuControl_minWidth"
-            backgroundStyle="primary"
             hoverBackgroundStyle="danger"
         />
     </div>
