+++ Controls-demo/Menu/Control/ItemActions/Index.wml
@@ -15,12 +15,4 @@
         </Router.router:Reference>
         <Controls-demo.Menu.Control.ItemActions.Menu.Index/>
     </div>
-    <div class="controlsDemo__cell">
-        <Router.router:Reference attr:class="controls-text-label"
-                                 state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/Menu/Control/ItemActions/MenuBackgroundStyle/Index' }}">
-            <a href="{{content.href}}">ItemActions, backgroundStyle = primary, hoverBackgroundStyle = danger</a>
-        </Router.router:Reference>
-        <Controls-demo.Menu.Control.ItemActions.MenuBackgroundStyle.Index/>
-    </div>
 </div>
