+++ Controls-demo/dropdown_new/Button/MenuPopupBackground/Index.wml
@@ -9,7 +9,7 @@
         </Controls.dropdown:Button>
     </div>
     <div class="controlsDemo__wrapper">
-        <div class="controls-text-label">backgroundStyle=primary, hoverBackgroundStyle = danger</div>
+        <div class="controls-text-label">popupBackgroundStyle=custom</div>
         <Controls.dropdown:Button
             keyProperty="key"
             caption="Create"
