+++ Controls/_menu/Control/Control.wml
@@ -1,6 +1,6 @@
 <Controls.scroll:ScrollContextProvider pagingVisible="{{ false }}">
    <Controls.dataSource:error.Container viewConfig="{{_errorConfig}}">
-      <div class="controls_breadcrumbs_theme-{{_options.theme}} controls_dropdownPopup_theme-{{_options.theme}} controls-menu controls-menu_backgroundStyle-{{_options.backgroundStyle}}"
+      <div class="controls_breadcrumbs_theme-{{_options.theme}} controls_dropdownPopup_theme-{{_options.theme}} controls-menu"
            attr:data-index="{{_options.subMenuLevel || 0}}" attr:ws-no-focus="true">
          <ws:if data="{{_listModel}}">
             <Controls.scroll:Container attr:class="controls-ScrollContainer__flex
