+++ Controls/_menu/Control/_Control.less
@@ -350,38 +350,6 @@
    margin-left: var(--grouping_padding-horizontal_menu);
 }
 
-.controls-menu_backgroundStyle-default {
-    background-color: var(--background-color);
-}
-
-.controls-menu_backgroundStyle-primary {
-    background-color: var(--primary_background-color);
-}
-
-.controls-menu_backgroundStyle-secondary {
-    background-color: var(--secondary_background-color);
-}
-
-.controls-menu_backgroundStyle-danger {
-    background-color: var(--danger_background-color);
-}
-
-.controls-menu_backgroundStyle-success {
-    background-color: var(--success_background-color);
-}
-
-.controls-menu_backgroundStyle-warning {
-    background-color: var(--warning_background-color);
-}
-
-.controls-menu_backgroundStyle-info {
-    background-color: var(--info_background-color);
-}
-
-.controls-menu_backgroundStyle-unaccented {
-    background-color: var(--unaccented_contrast_background-color);
-}
-
 @padding-right_menu-multiSelect: calc(var(--apply-button_inline-height_menuPopup) + var(--offset_xs) * 2);
 
 .controls-ListView__item-rightPadding_menu-multiselect {
@@ -436,35 +404,31 @@
         color: var(--pin-button_hover_icon-color_menu);
     }
 
-    .controls-menu_backgroundStyle-hover-default:hover {
-        background-color: var(--hover_background-color);
+    .controls-Menu__row_hoverBackgroundStyle-default:hover {
+        background-color: var(--item_hover_background-color_menu);
     }
 
-    .controls-menu_backgroundStyle-hover-primary:hover {
+    .controls-Menu__row_hoverBackgroundStyle-primary:hover {
         background-color: var(--primary_hover_background-color);
     }
 
-    .controls-menu_backgroundStyle-hover-secondary:hover {
+    .controls-Menu__row_hoverBackgroundStyle-secondary:hover {
         background-color: var(--secondary_hover_background-color);
     }
 
-    .controls-menu_backgroundStyle-hover-danger:hover {
+    .controls-Menu__row_hoverBackgroundStyle-danger:hover {
         background-color: var(--danger_hover_background-color);
     }
 
-    .controls-menu_backgroundStyle-hover-success:hover {
-        background-color: var(--success_hover_background-color);
-    }
-
-    .controls-menu_backgroundStyle-hover-warning:hover {
+    .controls-Menu__row_hoverBackgroundStyle-warning:hover {
         background-color: var(--warning_hover_background-color);
     }
 
-    .controls-menu_backgroundStyle-hover-info:hover {
+    .controls-Menu__row_hoverBackgroundStyle-info:hover {
         background-color: var(--info_hover_background-color);
     }
 
-    .controls-menu_backgroundStyle-hover-unaccented:hover {
+    .controls-Menu__row_hoverBackgroundStyle-unaccented:hover {
         background-color: var(--unaccented_hover_contrast_background-color);
     }
 }
