+++ Controls/_menu/Popup/_Popup.less
@@ -30,10 +30,6 @@
     background-color: var(--danger_background-color);
 }
 
-.controls-Menu__popup-background-success {
-    background-color: var(--success_background-color);
-}
-
 .controls-Menu__popup-background-warning {
     background-color: var(--warning_background-color);
 }
