+++ Controls/_menu/Popup/template.wml
@@ -68,6 +68,7 @@
                        selectorTemplate="{{_options.selectorTemplate}}"
                        isCompoundTemplate="{{_options.isCompoundTemplate}}"
                        itemActions="{{_options.itemActions}}"
+                       hoverBackgroundStyle="{{_options.hoverBackgroundStyle}}"
                        itemActionVisibilityCallback="{{_options.itemActionVisibilityCallback}}"
                        keyProperty="{{_options.keyProperty}}"
                        displayProperty="{{_options.displayProperty}}"
@@ -99,7 +100,7 @@
                        dataLoadErrback="{{_dataLoadErrback}}"
                        headingIcon="{{_headingIcon}}"
                        headerTemplate="{{_headerTemplate || _options.headerTemplate}}"
-                       hoverBackgroundStyle="{{_options.hoverBackgroundStyle}}"
+                       backgroundStyle="{{_options.backgroundStyle}}"
                        trigger="{{_options.trigger}}"
                        scrollbarVisible="{{_options.scrollbarVisible}}"
                        subMenuLevel="{{_options.subMenuLevel}}"
