+++ Controls/_menu/Render/itemTemplate.wml
@@ -1,6 +1,5 @@
 <ws:template name="menuRowTemplate">
-<div attr:class="controls-menu_backgroundStyle-hover-{{_options.hoverBackgroundStyle}}
-                 controls-Menu__row
+<div attr:class="controls-Menu__row
                  controls-ListView__itemV js-controls-ListView__measurableContainer
                  {{itemData.itemClassList}}
                  controls-Menu__row_{{multiLine ? 'multiLine' : 'singleLine'}}">
@@ -80,7 +79,7 @@
             <ws:partial template="{{ swipeTemplate }}"/>
          </ws:if>
          <ws:else>
-            <ws:partial template="{{ itemActionsTemplate }}" />
+            <ws:partial template="{{ itemActionsTemplate }}" style="{{_options.hoverBackgroundStyle}}"/>
          </ws:else>
       </ws:if>
    </ws:else>
